feat(scene-engine): add StoryStep with Block layout, move streaming infra from story feature
