+++ frontend/src/features/scene-engine/steps.ts
@@ -1,8 +1,9 @@
 import type { StepDef } from "./types";
 import PlaceholderStep from "./steps/PlaceholderStep";
+import StoryStep from "./steps/story/StoryStep";
 
 export const SCENE_STEPS: StepDef[] = [
-  { id: 1, label: "Story",                component: PlaceholderStep },
+  { id: 1, label: "Story",                component: StoryStep },
   { id: 2, label: "Character Extraction", component: PlaceholderStep },
   { id: 3, label: "Character Rendering",  component: PlaceholderStep },
   { id: 4, label: "Scene Extraction",     component: PlaceholderStep },
