+++ frontend/src/features/scene-engine/steps/story/story.types.ts
@@ -1,0 +1,21 @@
+export type StoryDetailType = {
+  id: string;
+  projectId: string;
+  storyText: string;
+  userInputText: string;
+  meta: Record<string, unknown>;
+  sourceEventId?: string;
+  error?: string;
+};
+
+export type EditEventType = {
+  id: string;
+  projectId: string;
+  targetType: string;
+  targetId: string;
+  operationType: string;
+  userInstruction: string;
+  outputSnapshot: { storyText?: string } | null;
+  status: string;
+  createdAt: string;
+};
