+++ frontend/src/features/story/phases/story-phase/hooks/useStoryStream.ts
@@ -1,38 +1,1 @@
-import type { EventEnvelope } from "@/features/story/events/base-events";
-import { httpClient } from "@/lib/httpClient";
-import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { useRef } from "react";
-import EventRouter from "../../../events/event-router";
-import { storyDetailOptions, storyHistoryOptions } from "../story-phase.queries";
-
-const STREAM_ENDPOINT = (projectId: string, storyId: string) =>
-  `/api/comic-builder/v2/project/${projectId}/story/${storyId}/generate` as const;
-
-export const useStoryStream = (projectId: string, storyId: string) => {
-  const queryClient = useQueryClient();
-  const queryKey = storyDetailOptions(projectId, storyId).queryKey;
-  const eventRouter = useRef(new EventRouter(queryClient, queryKey));
-
-  const mutation = useMutation({
-    mutationFn: (userInputText: string) =>
-      startGeneration(userInputText, storyId),
-    onSuccess: () => {
-      void queryClient.invalidateQueries({
-        queryKey: storyHistoryOptions(projectId, storyId).queryKey,
-      });
-    },
-  });
-
-  async function startGeneration(userInputText: string, storyId: string) {
-    const stream = httpClient.streamPost<EventEnvelope>(
-      STREAM_ENDPOINT(projectId, storyId),
-      { storyPrompt: userInputText },
-    );
-
-    for await (const event of stream) {
-      eventRouter.current.handle(event);
-    }
-  }
-
-  return mutation;
-};
+export { useStoryStream } from "@/features/scene-engine/steps/story/hooks/useStoryStream";
