+++ frontend/src/features/story/phases/story-phase/story-phase.types.ts
@@ -1,26 +1,9 @@
-export type StoryDetailType = {
-  id: string;
-  projectId: string;
-  storyText: string;
-  userInputText: string;
-  meta: Record<string, unknown>;
-  sourceEventId?: string;
-  error?: string;
-};
+export type {
+  StoryDetailType,
+  EditEventType,
+} from "@/features/scene-engine/steps/story/story.types";
 
 export type StoryStreamChunk = {
   delta?: string;
   finishReason?: string;
-};
-
-export type EditEventType = {
-  id: string;
-  projectId: string;
-  targetType: string;
-  targetId: string;
-  operationType: string;
-  userInstruction: string;
-  outputSnapshot: { storyText?: string } | null;
-  status: string;
-  createdAt: string;
 };
